requirements/base.txt
django==2.1
djangorestframework==3.8.2
django-choices==1.6.1
django-filter==2.0.0
django-api-key==0.1.0
coreapi==2.3.3
pygments==2.2.0
markdown==2.6.11
PyMySQL==0.9.2
gunicorn==19.9.0
pillow==5.3.0
redis==2.10.6
pycrc==1.21
semantic_version==2.6.0
celery==4.2.1
django-redis==4.10.0
